Adding fuzzy functionality
Putting everything in place for a basic fuzzy logic controller for the AI paddle. Right now I just want it to hit the ball, will focus on having it use a strategy (like in the final MATLAB version) after I get this initial one working.

The fuzzy system in this needs to be finished, and the input membership functions for fuzzypaddle.move_fuzzy() need to be converted for playing on a game that's 275 units tall, not 100.