Very basic fuzzy paddle, buggy
Finally got a very basic version of the fuzzy paddle going. Only problem: it runs SUUUUPER slow!! So I need to figure out why that's happening.

Also I get an error that the index for the movement output is out of bounds, and I think I need to make the output mfs in pdlemvmnt have a range of [0:2] and then convert movement back to [-1:1] after. I think the negatives might be screwing with this toolbox